+++ Nagisa/MainPage.xaml.h
@@ -6,11 +6,14 @@
 #pragma once
 
 #include "MainPage.g.h"
+#include "XamlResources\CustomConverters.xaml.h"
 
 namespace Nagisa
 {
 	using Assassin::TransferManager;
+	using Assassin::TransferTask;
 	using Platform::Object;
+	using Platform::Collections::Vector;
 	using Windows::UI::Xaml::RoutedEventArgs;
 
 	public ref class MainPage sealed
@@ -22,5 +25,7 @@ namespace Nagisa
 	private:
 		void AboutButtonClick(Object^ sender, RoutedEventArgs^ e);
 		void NewTaskButtonClick(Object^ sender, RoutedEventArgs^ e);
+		void Page_Loaded(Object^ sender, RoutedEventArgs^ e);
+		void CopyLinkMenuItem_Click(Object^ sender, RoutedEventArgs^ e);
 	};
 }